(feat): Publish to NPM (alpha), (beta), and (latest)
This commit updates the packagePublish.yml workflow to include separate steps for publishing to NPM based on the version tag. It introduces three new steps: Publish to npm (alpha), Publish to npm (beta), and Publish to npm (latest). The step Publish to npm (alpha) is triggered when the version contains '-alpha', and it publishes the package with the 'alpha' tag. Similarly, the step Publish to npm (beta) is triggered when the version contains '-beta', and it publishes the package with the 'beta' tag. Finally, the step Publish to npm (latest) is triggered when the version does not contain '-alpha' or '-beta', and it publishes the package without any tag. This change improves the release process by allowing different versions of the package to be published with appropriate tags.

Co-authored-by: Lobo Metalúrgico <43734867+LoboMetalurgico@users.noreply.github.com>